Layers

Simple Layers

NeuralVerifier.Encoding.dense — Function

dense(x, W, b)

Fully-connected or dense layer that given a weight vector or matrix W, and a bias b, compute the linear function y = Wx + b

NeuralVerifier.Encoding.flatten — Function

flatten(x)

Flatten a multidimensional matrix into a vector

Convolutional Layers

NeuralVerifier.Encoding.conv1D — Function

conv1D(x::Array{PyObject,1}, filter, stride_size = 2)

Apply a 1D convolution to x using the filter matrix.

NeuralVerifier.Encoding.conv2D — Function

conv2D(x::Array{PyObject,2}, filter, stride_size = (2,2))

Apply a 2D Convolutional operation to a 2D matrix x, using the filter matrix as the weight matrices.

Pooling Layers

NeuralVerifier.Encoding.maxpool — Function

maxpool(x; poolsize=2, stride=2)

Max pooling operation.

NeuralVerifier.Encoding.avgpool — Function

avgpool(x; poolsize = 2, stride=2)

Average pooling operation